+++ src/TestSuite/Traits/TestCaseTrait.php
@@ -16,7 +16,6 @@ namespace MeTools\TestSuite\Traits;
 use Cake\Core\Plugin;
 use Cake\Filesystem\Folder;
 use Exception;
-use PHPUnit\Framework\MockObject\MockObject;
 use Tools\ReflectionTrait;
 use Tools\TestSuite\TestCaseTrait as ToolsTestCaseTrait;
 
@@ -64,19 +63,6 @@ trait TestCaseTrait
         }
 
         $this->assertContains($expectedContent, file_get_contents($filename), $message);
-    }
-
-    /**
-     * Asserts that an object is a mock
-     * @param object $object Object
-     * @param string $message The failure message that will be appended to the
-     *  generated message
-     * @return void
-     * @since 2.17.5
-     */
-    public function assertIsMock($object, $message = '')
-    {
-        $this->assertInstanceOf(MockObject::class, $object, $message);
     }
 
     /**