+++ CHANGELOG.md
@@ -15,7 +15,8 @@
     only suggested packages;
 * removed `ME_TOOLS` constants. It no longer uses also the `ASSETS` constant;
 * removed `clearDir()` global function. Use instead `safe_unlink_recursive()`
-    provided by `php-tools`.
+    provided by `php-tools`;
+* removed `MockTrait::assertIsMock()`.
 
 ## 2.17 branch
 ### 2.17.5